// src/app/shared/models/VM/basicData/PeriodVM.ts
export class PeriodVM{
    id: number;
    periodFrom: string;
    periodTo: string;
    periodOfTheDayFrom: string;
    periodOfTheDayFromAr: string;
    periodOfTheDayTo: string;
    periodOfTheDayToAr: string;

    createdBy: string;
    createdDate: string;
    updatedBy: string;
    updatedDate: string;
}